refactor: remove unused interface
This removes the `IStrategyScoreMethod` and `INextWordGuesserStrategyBase` interfaces.

I initially planned to have multiple solver classes implementing the same pluggable interface, but I realized that most of the functionality to be shared would be the same; so it is better to extend a core abstract class instead, just differentiating on that unique `scoreForGuess` method.